Break up autosort and comment directives
